requirement.txt
absl-py==0.7.1
astor==0.7.1
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2019.3.9
cycler==0.10.0
Cython==0.29.6
easydict==1.9
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
html5lib==0.9999999
imutils==0.5.2
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.1.0
Markdown==3.1
matplotlib==2.2.4
mbutil==0.3.0
mock==2.0.0
numpy==1.16.2
opencv-python==3.2.0.8
pbr==5.1.3
Pillow==6.0.0
pkg-resources==0.0.0
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
scikit-learn==0.20.3
scipy==1.2.1
Shapely==1.6.4.post2
six==1.12.0
sklearn==0.0
subprocess32==3.5.3
tensorboard==1.6.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
Werkzeug==0.15.2
